python/requirements.txt
black==23.7.0
click==8.1.6
kafka-python==2.0.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
tomli==2.0.1
typing-extensions==4.7.1
